% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shop.R
\name{eats}
\alias{eats}
\title{what would you like to eat}
\usage{
eats(what)
}
\arguments{
\item{eat}{shop name.}
}
\value{
shop name
}
\description{
choose randomly shop name
}
\examples{
## a trivial example
eats(what)
}

