% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shop.R
\name{score}
\alias{score}
\title{Simulate google map score}
\usage{
score(name, n)
}
\arguments{
\item{name}{a shop name.}
\item{n}{an integer.}
}
\value{
a score after simulation.
}
\description{
Simulate google map score
}
\examples{
## a trivial example
score("The Old Factory Restaurant", 30)
}

